Practica MVC

+++ Peliculas/src/peliculas/datos/PeliculaDAO.java
@@ -1,30 +1,115 @@
 package peliculas.datos;
 
 import peliculas.domain.Pelicula;
+import peliculas.excepciones.AccesoDatosEx;
+import peliculas.excepciones.EscrituraDatosEx;
+import peliculas.excepciones.LecturaDatosEx;
 
 import java.io.*;
-import java.util.List;
+import java.util.*;
 
 public class PeliculaDAO implements IAccesoDatos
 {
-    @Override
-    public boolean existe(String nombreArchivo)
+    public PeliculaDAO()
     {
-        File archivo = new File(nombreArchivo);
+
+    }
+    /*
+    Metodo para verificar si existe un archivo
+     */
+    @Override
+    public boolean existe(String recurso)
+    {
+        File archivo = new File(recurso);
         return archivo.exists();
     }
 
+    /*
+    Metodo para leer todas las peliculas
+    Retorna un listado de Peliculas
+     */
     @Override
-    public List<Pelicula> listar(String nombre)
+    public List<Pelicula> listar(String recurso) throws LecturaDatosEx
     {
-        File archivo = new File(nombre);
+        File archivo = new File(recurso);
+        List<Pelicula> peliculas = new ArrayList<Pelicula>();
+
+        try
+        {
+            BufferedReader entrada = new BufferedReader(new FileReader(archivo));
+            String lectura = entrada.readLine();
+
+            while(lectura != null)
+            {
+                peliculas.add(new Pelicula(lectura));
+                lectura = entrada.readLine();
+            }
+
+            entrada.close();
+            System.out.println("Se obtiene listado de peliculas");
+        }
+        catch (FileNotFoundException e)
+        {
+            e.printStackTrace(System.out);
+            throw new LecturaDatosEx("Excepcion al listar peliculas " + e.getMessage());
+        }
+        catch (IOException e)
+        {
+            e.printStackTrace(System.out);
+            throw new LecturaDatosEx("Excepcion al listar peliculas " + e.getMessage());
+        }
+        return peliculas;
+    }
+
+    /*
+    Metodo para escribir una pelicula en un archivo determinado
+     */
+    @Override
+    public void escribir(Pelicula pelicula, String recurso, boolean anexar) throws EscrituraDatosEx
+    {
+        File archivo = new File(recurso);
+        try
+        {
+            PrintWriter salida = new PrintWriter(new FileWriter(archivo,anexar));
+            salida.println(pelicula.getNombre());
+            salida.close();
+            System.out.println("Se ha escrito en el archivo");
+        }
+        catch (FileNotFoundException e)
+        {
+            e.printStackTrace(System.out);
+            throw new EscrituraDatosEx("Excepcion al escribir pelicula " + e.getMessage());
+        }
+        catch (IOException e)
+        {
+            e.printStackTrace(System.out);
+            throw new EscrituraDatosEx("Excepcion al escribir pelicula " + e.getMessage());
+        }
+
+    }
+
+    /*
+    Metodo para buscar un nombre de pelicula en el archivo indicado
+     */
+    @Override
+    public String buscar(String recurso, String buscar) throws LecturaDatosEx
+    {
+        File archivo = new File(recurso);
+        String resultado = null;
+        int indice = 1;
+
         try
         {
             BufferedReader entrada = new BufferedReader(new FileReader(archivo));
             String lectura = entrada.readLine();
             while(lectura != null)
             {
-                System.out.println(lectura);
+                if(buscar != null && buscar.equalsIgnoreCase(lectura))
+                {
+                    resultado = "Pelicula " + lectura + " encontrada en el indice " + indice;
+                    break;
+                }
+                indice++;
                 lectura = entrada.readLine();
             }
             entrada.close();
@@ -32,86 +117,53 @@ public class PeliculaDAO implements IAccesoDatos
         catch (FileNotFoundException e)
         {
             e.printStackTrace(System.out);
+            throw new LecturaDatosEx("Excepcion al buscar pelicula "+ e.getMessage());
         }
         catch (IOException e)
         {
             e.printStackTrace(System.out);
+            throw new LecturaDatosEx("Excepcion al buscar pelicula "+ e.getMessage());
         }
-
-        return null;
+        return resultado;
     }
 
+    /*
+    Metodo para crear un archivo vacio
+     */
     @Override
-    public void escribir(Pelicula pelicula, String nombreArchivo, boolean anexar)
+    public void crear(String recurso) throws AccesoDatosEx
     {
-        File archivo = new File(nombreArchivo);
+        File archivo = new File(recurso);
         try
         {
-            PrintWriter salida = new PrintWriter(new FileWriter(archivo,anexar));
-            salida.println(pelicula);
-            salida.close();
-            System.out.println("Se ha escrito en el archivo");
-        }
-        catch (FileNotFoundException e)
-        {
-            e.printStackTrace(System.out);
-        }
-        catch (IOException e)
-        {
-            e.printStackTrace(System.out);
-        }
-
-    }
-
-    @Override
-    public String buscar(String nombreArchivo, String buscar)
-    {
-        File archivo = new File(nombreArchivo);
-        String nombreBuscado = null;
-        try
-        {
-            BufferedReader entrada = new BufferedReader(new FileReader(archivo));
-            String lectura = entrada.readLine();
-            while(lectura.equals(buscar))
-            {
-                lectura = entrada.readLine();
-
-            }
-            nombreBuscado = lectura;
-        }
-        catch (FileNotFoundException e)
-        {
-            e.printStackTrace(System.out);
-        }
-        catch (IOException e)
-        {
-            e.printStackTrace(System.out);
-        }
-        return nombreBuscado;
-    }
-
-    @Override
-    public void crear(String nombreArchivo)
-    {
-        File archivo = new File(nombreArchivo);
-        try
-        {
-            PrintWriter salida = new PrintWriter(archivo);
+            PrintWriter salida = new PrintWriter(new FileWriter(archivo));
             salida.close();
             System.out.println("Se ha creado el archivo");
         }
         catch (FileNotFoundException e)
         {
             e.printStackTrace(System.out);
+            throw new LecturaDatosEx("Excepcion al crear archivo "+ e.getMessage());
+        }
+        catch (IOException e)
+        {
+            e.printStackTrace();
+            throw new LecturaDatosEx("Excepcion al crear archivo "+ e.getMessage());
         }
     }
 
+    /*
+    Metodo para borrar un archivo
+     */
     @Override
-    public void borrar(String nombreArchivo)
+    public void borrar(String recurso)
     {
-        File archivo = new File(nombreArchivo);
+        File archivo = new File(recurso);
 
-        archivo.delete();
-
+        if(archivo.exists())
+        {
+            archivo.delete();
+            System.out.println("Se ha borrado el archivo");
+        }
     }
 }
